Refactor(LottoViewModel): 뷰 모델 로직 수정

+++ app/src/main/java/com/geniusjun/lotto/model/LottoUiState.kt
@@ -1,7 +1,7 @@
 package com.geniusjun.lotto.model
 
 data class LottoUiState(
-    val balance: Int,
+    val balance: Long,
     val winningNumbers: List<Int>,
     val bonusNumber: Int = 0
 )
